feat: add dependency injection

+++ internal/services/link/count.go
@@ -6,10 +6,10 @@ import (
 	"github.com/h-varmazyar/p3o/internal/errors"
 )
 
-func (s Service)TotalLinkCount(ctx context.Context, userId uint) (uint, error) {
+func (s Service) TotalLinkCount(ctx context.Context, userId uint) (uint, error) {
 	if count, err := s.linkRepo.TotalLinkCount(ctx, userId); err != nil {
 		return 0, errors.ErrUnexpected.AddOriginalError(err)
 	} else {
-		return count, nil
+		return uint(count), nil
 	}
 }
